requirements.txt
mysql-connector-python==8.0.28
pyTelegramBotAPI==4.4.0
python-dotenv==0.20.0
